+++ src/components/settings/sections/index.ts
@@ -1,4 +1,5 @@
 export { DependenciesSection } from './DependenciesSection';
 export { DownloadSection } from './DownloadSection';
+export { ExtensionSection } from './ExtensionSection';
 export { GeneralSection } from './GeneralSection';
 export { NetworkSection } from './NetworkSection';